Revert "Feature inventory page"

+++ purpure-busness/src/context/ProductsContext.tsx
@@ -1,18 +1,12 @@
-import { createContext, useState } from "react";
-import { AxiosError } from "axios";
-
+import { createContext, useEffect, useState } from "react";
 import { toast } from "react-toastify";
 import api from "../services/api";
-
-interface iApiError {
-  error: string;
-}
 
 export interface iProducts {
   product_name: string;
   product_value: string;
   product_stock: string;
-  userId: number | null;
+  userId: number;
   id?: number;
 }
 
@@ -22,45 +16,41 @@ interface iProductProps {
 
 interface iProductContext {
   products: iProducts[];
-  setProducts: React.Dispatch<React.SetStateAction<iProducts[]>>;
-  loadingClientProducts: () => void;
-  registerProduct: (data: iProducts) => Promise<void>;
+  registerProduct: (data: iProducts) => void;
   deleteProduct: (deletedProduct: iProducts) => void;
-  editProduct: (editedProduct: iProducts, productId: number | null) => void;
-  styleModal: boolean;
-  setStyleModal: React.Dispatch<React.SetStateAction<boolean>>;
+  editProduct: (editedProduct: iProducts) => void;
 }
 
 export const ProductContext = createContext({} as iProductContext);
 
 const ProductPovider = ({ children }: iProductProps) => {
   const [products, setProducts] = useState([] as iProducts[]);
-  const [styleModal, setStyleModal] = useState(false)
 
-  const loadingClientProducts = async () => {
-    const token = localStorage.getItem("@accessToken");
-    setStyleModal(true)
-    if (token) {
-      try {
-        const id = localStorage.getItem("@USER_ID");
+  useEffect(() => {
+    async function loadingClients() {
+      const token = localStorage.getItem("@accessToken");
+      const id = localStorage.getItem("@USER_ID");
 
-        api.defaults.headers.authorization = `Bearer ${token}`;
-        const { data } = await api.get(`/users/${id}?_embed=products`);
-        setProducts(data.products);
-      } catch (error) {
-        const requestError = error as AxiosError<iApiError>;
-        toast.error(requestError?.request.data.error);
-        console.log(error);
+      if (token) {
+        try {
+          api.defaults.headers.authorization = `Bearer ${token}`;
+          const { data } = await api.get(`/users/${id}?_embed=products`);
+          setProducts(data.products);
+        } catch (error) {
+          console.log(error);
+        }
       }
     }
-  };
+    loadingClients();
+  }, []);
 
   const registerProduct = async (data: iProducts) => {
-    console.log(data)
     if (
       !products.find((product) => product.product_name === data.product_name)
     ) {
       try {
+        const token = localStorage.getItem("@accessToken");
+
         const newProduct = [
           ...products,
           {
@@ -70,15 +60,14 @@ const ProductPovider = ({ children }: iProductProps) => {
             userId: data.userId,
           },
         ];
-        const token = localStorage.getItem("@accessToken");
+
         api.defaults.headers.authorization = `Bearer ${token}`;
-        console.log(data);
         await api.post("/products", data);
-        toast.success("Produto cadastrado com sucesso!");
+
+        toast.success("Cliente cadastrado com sucesso!");
+
         setProducts(newProduct);
       } catch (error) {
-        const requestError = error as AxiosError<iApiError>;
-        toast.error(requestError?.request.data.error);
         console.log(error);
       }
     } else {
@@ -88,58 +77,45 @@ const ProductPovider = ({ children }: iProductProps) => {
 
   const deleteProduct = async (deletedProduct: iProducts) => {
     // eslint-disable-next-line no-restricted-globals
-    console.log(deletedProduct);
-    try {
-      const newProductList = products.filter(
-        (product) => product.id !== deletedProduct.id
-      );
-      const token = localStorage.getItem("@accessToken");
-      api.defaults.headers.authorization = `Bearer ${token}`;
-      await api.delete(`/products/${deletedProduct.id}`);
-      setProducts(newProductList);
-      toast.success("O produto foi apagado da sua lista!");
-    } catch (error) {
-      const requestError = error as AxiosError<iApiError>;
-      
-        toast.error(requestError?.request.data.error);
+    if (confirm("Deseja excluir este produto?")) {
+      try {
+        const newProductList = products.filter(
+          (product) => product.id !== deletedProduct.id
+        );
+        const token = localStorage.getItem("@accessToken");
+
+        api.defaults.headers.authorization = `Bearer ${token}`;
+        await api.delete(`/products/${deletedProduct.id}`);
+        setProducts(newProductList);
+        toast.success("O produto foi apagado da sua lista!");
+      } catch (error) {
         console.log(error);
+      }
     }
   };
 
-  const editProduct = async (
-    editedProduct: iProducts,
-    productId: number | null
-  ) => {
+  const editProduct = async (editedProduct: iProducts) => {
     try {
+      const token = localStorage.getItem("@accessToken");
+
       const pachProduct = {
         product_name: editedProduct.product_name,
         product_value: editedProduct.product_value,
         product_stock: editedProduct.product_stock,
       };
-      const token = localStorage.getItem("@accessToken");
+
       api.defaults.headers.authorization = `Bearer ${token}`;
-      await api.patch(`/products/${productId}`, pachProduct);
+      await api.patch(`/products/${editedProduct.id}`, pachProduct);
 
       toast.success("O produto foi editado com sucesso!");
     } catch (error) {
-      const requestError = error as AxiosError<iApiError>;
-      toast.error(requestError?.request.data.error);
       console.log(error);
     }
   };
 
   return (
     <ProductContext.Provider
-      value={{
-        products,
-        setProducts,
-        loadingClientProducts,
-        registerProduct,
-        deleteProduct,
-        editProduct,
-        styleModal, 
-        setStyleModal,
-      }}
+      value={{ products, registerProduct, deleteProduct, editProduct }}
     >
       {children}
     </ProductContext.Provider>
